+++ lib/src/discrete_mathematics/linear_algebra/matrix/square_matrix.dart
@@ -11,12 +11,10 @@ class SquareMatrix extends Matrix {
   /// Constructor accept array of arrays of num numbers
   ///
   /// Count of inner arrays must be equal to count of their elements!
-  SquareMatrix(List<List<num>> data) : super(<List<num>>[]) {
+  SquareMatrix(List<List<num>> data) : super(data) {
     if (data.length != data[0].length) {
       throw MatrixException(
           'Count of matrix columns must be equal to count of rows!');
-    } else {
-      this.data = data;
     }
   }
 
@@ -40,10 +38,12 @@ class SquareMatrix extends Matrix {
       return itemAt(1, 1);
     } else {
       num res = 0;
-      for (var i = 0; i < firstRow.length; i++) {
-        final changedMatrix =
-            SquareMatrix(removeRow(1).removeColumn(i + 1).data);
-        res += pow(-1, 1 + (i + 1)) * firstRow[i] * changedMatrix.determinant();
+      for (var i = 1; i <= columns; i++) {
+        final changedMatrix = SquareMatrix(data);
+        changedMatrix
+          ..removeRow(1)
+          ..removeColumn(i);
+        res += pow(-1, 1 + i) * firstRow[i - 1] * changedMatrix.determinant();
       }
       return res;
     }
@@ -52,7 +52,7 @@ class SquareMatrix extends Matrix {
   /// Checks if this matrix is singular
   bool isSingular() => determinant() == 0;
 
-  /// Checks if this matrix is singular
+  /// Checks if this matrix isn't singular
   bool isNotSingular() => !isSingular();
 
   /// Checks if this matrix is symmetric
@@ -93,8 +93,10 @@ class SquareMatrix extends Matrix {
       final matrixOfCofactors = SquareMatrix.generate(rows);
       for (var i = 1; i <= rows; i++) {
         for (var j = 1; j <= columns; j++) {
-          final nData = origin.removeRow(i).removeColumn(j).data;
-          final det = SquareMatrix(nData).determinant();
+          origin
+            ..removeRow(i)
+            ..removeColumn(j);
+          final det = origin.determinant();
           matrixOfCofactors.setItem(i, j, pow(-1, i + j) * det);
         }
       }
@@ -139,45 +141,64 @@ class SquareMatrix extends Matrix {
   /// Singular value decomposition for this matrix
   // Map<String, Vector> svd() {}
 
-  /// Performs Gaussian elimination of this matrix and [equalTo] as right-side of augmented matrix
+  /// Performs Gaussian-Jordan elimination of this matrix and [equalTo] as right-side of augmented matrix
   ///
   /// [equalTo] should be equal to [rows].
-  Vector eliminate(List<num> equalTo) {
-    var matrix = SquareMatrix(data);
-    final vector = equalTo;
-
-    while (itemAt(1, 1) == 0) {
-      final row = matrix.rowAt(1);
-      final value = vector.removeAt(0);
-      matrix = matrix.removeRow(1).insertRow(row).toSquareMatrix();
-      vector.add(value);
-    }
+  Vector eliminate(List<double> equalTo) {
+    final eliminatedMatrix = SquareMatrix(data);
+    final result = equalTo;
 
     for (var i = 1; i <= rows; i++) {
-      final element = itemAt(i, i);
-      final row1 = matrix.rowAt(i).map((v) => v / element).toList();
-      vector[i - 1] /= element;
-      matrix = matrix.insertRow(row1, index: i).toSquareMatrix();
+      if (eliminatedMatrix.itemAt(i, i) == 0) {
+        final col = eliminatedMatrix.columnAt(i);
+        final row = eliminatedMatrix.rowAt(i);
 
-      for (var j = i + 1; j <= columns; j++) {
-        final tmpRow = Vector(row1).transform((v) => v * (-1) * itemAt(j, i)) +
-            Vector(rowAt(j));
-        vector[j - 1] += vector[i - 1] * (-1) * itemAt(j, i);
-        matrix = matrix.insertRow(tmpRow.data, index: j).toSquareMatrix();
+        final indexCol = row.indexWhere((n) => n != 0, i - 1);
+        final indexRow = col.indexWhere((n) => n != 0, i - 1);
+
+        if (indexRow != -1) {
+          eliminatedMatrix.swapRows(i, indexRow + 1);
+
+          final value = result[i - 1];
+          result[i - 1] = result[indexRow];
+          result[indexRow] = value;
+        } else if (indexCol != -1) {
+          eliminatedMatrix.swapColumns(i, indexCol + 1);
+        }
+      }
+
+      final element = eliminatedMatrix.itemAt(i, i) == 0
+          ? 1
+          : eliminatedMatrix.itemAt(i, i);
+      final choosedRow =
+          eliminatedMatrix.rowAt(i).map((v) => v / element).toList();
+      eliminatedMatrix.replaceRow(i, choosedRow);
+      result[i - 1] /= element;
+
+      for (var j = i + 1; j <= rows; j++) {
+        final diff =
+            eliminatedMatrix.itemAt(j, i) / eliminatedMatrix.itemAt(i, i);
+        final tmpRow = Vector(choosedRow).transform((v) => v * -diff) +
+            Vector(eliminatedMatrix.rowAt(j));
+
+        result[j - 1] += result[i - 1] * -diff;
+        eliminatedMatrix.replaceRow(j, tmpRow.data);
       }
     }
 
     for (var i = rows; i >= 1; i--) {
-      final row1 = matrix.rowAt(i);
+      final choosedRow = eliminatedMatrix.rowAt(i);
 
       for (var j = i - 1; j >= 1; j--) {
-        final tmpRow = Vector(row1).transform((v) => v * (-1) * itemAt(j, i)) +
-            Vector(rowAt(j));
-        vector[j - 1] += vector[i - 1] * (-1) * itemAt(j, i);
-        matrix = matrix.insertRow(tmpRow.data, index: j).toSquareMatrix();
+        final tmpRow = Vector(choosedRow)
+                .transform((v) => v * -eliminatedMatrix.itemAt(j, i)) +
+            Vector(eliminatedMatrix.columnAt(j));
+
+        result[j - 1] += result[i - 1] * -eliminatedMatrix.itemAt(j, i);
+        eliminatedMatrix.replaceRow(j, tmpRow.data);
       }
     }
 
-    return Vector(vector);
+    return Vector(result);
   }
 }