+++ lib/src/discrete_mathematics/linear_algebra/vector/base/vector_base.dart
@@ -10,14 +10,17 @@ abstract class VectorBase {
   /// Default constructor that don't accept data
   VectorBase();
 
-  /// Constructor that accept [data]
-  VectorBase.init(this.data);
+  /// Constructor that accept [_data]
+  VectorBase.init(this._data);
 
   /// Data for vector
-  List<num> data;
+  List<num> _data;
+
+  /// Gets [_data] of this vector
+  List<num> get data => _data.toList();
 
   /// Count of vector's numbers
-  int get itemCount => data.length;
+  int get itemCount => _data.length;
 
   /// Gets length of this vector
   num get length => euclideanNorm();
@@ -25,13 +28,13 @@ abstract class VectorBase {
   /// Gets number at specified [index]
   ///
   /// [index] is in range from 1 to end inclusively.
-  num itemAt(int index) => data[index - 1];
+  num itemAt(int index) => _data[index - 1];
 
   /// Sets [item] in specified [position] of vector
   ///
   /// [position] should starts from 1.
   void setItem(int position, num item) {
-    data[position - 1] = item;
+    _data[position - 1] = item;
   }
 
   /// Gets norm of vector alse known as vector's length
@@ -43,7 +46,7 @@ abstract class VectorBase {
       throw VectorException('P must be greater or equal to 1! Given $p.');
     } else if (n.isInfinite) {
       var res = itemAt(1);
-      for (var i in data) {
+      for (var i in _data) {
         if (res < i) {
           res = i;
         }
@@ -51,7 +54,7 @@ abstract class VectorBase {
       return res;
     } else {
       var sum = 0.0;
-      for (var item in data) {
+      for (var item in _data) {
         sum += pow(item, n);
       }
       return pow(sum, 1 / n);
@@ -66,7 +69,7 @@ abstract class VectorBase {
 
   /// Multiply this vector by [vector] using dot product algorithm
   num dotProduct(covariant VectorBase vector) =>
-      toMatrix().transpose().multiplyByMatrix(vector.toMatrix()).itemAt(1, 1);
+      toMatrixRow().multiplyByMatrix(vector.toMatrixColumn()).itemAt(1, 1);
 
   /// Gets cross product of this vector and another [vector]
   ///
@@ -74,8 +77,11 @@ abstract class VectorBase {
   /// [itemCount] of both vectors must be equal to 3.
   VectorBase crossProduct(covariant VectorBase vector);
 
-  /// Convert this vector to martix
-  MatrixBase toMatrix();
+  /// Convert this vector to martix with one row
+  MatrixBase toMatrixRow();
+
+  /// Converts this vector to matrix with one column
+  MatrixBase toMatrixColumn();
 
   /// Get angle between this vector and another [vector]
   ///
@@ -109,24 +115,38 @@ abstract class VectorBase {
   bool isOrthonormalWith(covariant VectorBase vector) =>
       isOrthogonalTo(vector) && isUnit() && vector.isUnit();
 
-  /// Multiplies two vectors using Hadamard product algorithm
-  VectorBase operator *(covariant VectorBase vector) => hadamardProduct(vector);
+  /// Multiplies this vector by number or other vector
+  ///
+  /// When multiplying two vectors, Hadamard's product is used.
+  VectorBase operator *(Object other) {
+    VectorBase v;
+    if (other is num) {
+      v = transform((v) => v * other);
+    } else if (other is VectorBase) {
+      v = hadamardProduct(other);
+    }
+    return v;
+  }
 
   /// Divides corresponding values of this vactors by [vector]
-  VectorBase operator /(covariant VectorBase vector) =>
-      this * vector.transform((v) => 1 / v);
+  VectorBase operator /(num number) => this * 1 / number;
 
   /// Add values of this vector to [vector]'s values
   VectorBase operator +(covariant VectorBase vector);
 
+  /// Gets unary minus of this vector
+  VectorBase operator -() => transform((v) => -v);
+
   /// Subtract [vector]'s values from values of this vector
-  VectorBase operator -(covariant VectorBase vector) =>
-      this + vector.transform((v) => -v);
+  VectorBase operator -(covariant VectorBase vector) => this + -vector;
 
   @override
   bool operator ==(Object other) =>
       other is VectorBase && hashCode == other.hashCode;
 
   @override
-  int get hashCode => hashObjects(data);
+  int get hashCode => hashObjects(_data);
+
+  @override
+  String toString() => '$_data';
 }