+++ test/extended_math_test.dart
@@ -3,10 +3,21 @@ import 'dart:math';
 import 'package:extended_math/extended_math.dart';
 
 void main() {
-  final i1 = Complex(re: 3, im: 3.9);
-  final i2 = Complex(re: 3.1, im: 1);
+  final i1 = Matrix(<List<num>>[
+    <int>[1, 0, 0, 0],
+    <int>[0, 0, 0, 0],
+    <int>[1, -7, 0, 0]
+  ]);
 
-  final cn = CompositeNumber(25);
+  final i2 = SquareMatrix(<List<num>>[
+    <int>[3, 2, 2],
+    <int>[0, 3, 1],
+    <int>[0, 3, 6]
+  ]);
 
-  print(i1.argument);
+  final v1 = Vector(<int>[6, 5, 7]);
+  final v2 = Vector(<int>[6, 3, 1]);
+
+  print(i1.rowAt(1).removeLast());
+  print(i1);
 }